test: ヘルパーメソッドをモジュールに切り分けてヘルパースペックを作成 (#68)
