+++ app/helpers/user_sessions_helper.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+module UserSessionsHelper
+  def current_user
+    return unless session[:user_id]
+
+    @current_user ||= User.find(session[:user_id])
+  end
+
+  def logged_in?
+    !!current_user
+  end
+end
